index: add a dirty bit reflecting unsaved changes
Teach the index when it is "dirty", and has unsaved changes.  Consider
the index dirty whenever a caller has added or removed an entry from the
main index, REUC or NAME section.  Similarly, consider the index not
dirty immediately upon clear, read or write.

This allows us to ensure that unsaved changes are not lost when we
automatically refresh the index.